+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.engine.CarEngine;
+import racingcar.domain.car.engine.CarEngineFactory;
 
 public class CarTest {
     Car car;
@@ -16,28 +16,28 @@ public class CarTest {
 
     @Test
     public void 테스트_자동차움직임() {
-        car.move(CarEngine.forward());
+        car.move(CarEngineFactory.forward());
         assertEquals(car.getPosition().get(), 1);
 
-        car.move(CarEngine.forward());
+        car.move(CarEngineFactory.forward());
         assertEquals(car.getPosition().get(), 2);
 
-        car.move(CarEngine.forward());
+        car.move(CarEngineFactory.forward());
         assertEquals(car.getPosition().get(), 3);
 
-        car.move(CarEngine.stop());
+        car.move(CarEngineFactory.stop());
         assertEquals(car.getPosition().get(), 3);
     }
 
     @Test
     public void 테스트_현재위치표시() {
-        car.move(CarEngine.forward());
+        car.move(CarEngineFactory.forward());
         assertEquals(car.toString(), "pobi : -");
 
-        car.move(CarEngine.forward());
+        car.move(CarEngineFactory.forward());
         assertEquals(car.toString(), "pobi : --");
 
-        car.move(CarEngine.stop());
+        car.move(CarEngineFactory.stop());
         assertEquals(car.toString(), "pobi : --");
     }
 }
